Rearrange startup code
